requirements.txt
streamlit 
snowflake.connector
pandas 
numpy 
sklearn
tensorflow